new diagram rendering engin
